+++ lib/elasticsearch_ex/api/document.ex
@@ -132,6 +132,8 @@ defmodule ElasticsearchEx.Api.Document do
   Refer to the official [documentation](https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-get.html#docs-get-api-query-params)
   for a detailed list of the parameters.
 
+  You can provide the option `source_only: true` to return only the source of the document.
+
   ### Examples
 
       iex> ElasticsearchEx.Api.Document.get(index: "my-index-000001", id: "0")
@@ -155,25 +157,10 @@ defmodule ElasticsearchEx.Api.Document do
          "_version" => 1,
          "found" => true
        }}
-  """
-  @spec get(keyword()) :: ElasticsearchEx.response()
-  def get(opts \\ []) when is_list(opts) do
-    {index, document_id, opts} = extract_required_index_and_required_id!(opts)
 
-    Client.get("#{index}/_doc/#{document_id}", nil, nil, opts)
-  end
+  ### Examples - Returns only the source
 
-  @doc """
-  Retrieves the specified JSON source from an index.
-
-  ### Query parameters
-
-  Refer to the official [documentation](https://www.elastic.co/guide/en/elasticsearch/reference/current/docs-get.html#docs-get-api-query-params)
-  for a detailed list of the parameters.
-
-  ### Examples
-
-      iex> ElasticsearchEx.Api.Document.get_source(index: "my-index-000001", id: "0")
+      iex> ElasticsearchEx.Api.Document.get(index: "my-index-000001", id: "0", source_only: true)
       {:ok,
        %{
          "@timestamp" => "2099-11-15T14:12:12",
@@ -187,11 +174,13 @@ defmodule ElasticsearchEx.Api.Document do
          "user" => %{"id" => "kimchy"}
        }}
   """
-  @spec get_source(keyword()) :: ElasticsearchEx.response()
-  def get_source(opts \\ []) when is_list(opts) do
+  @spec get(keyword()) :: ElasticsearchEx.response()
+  def get(opts \\ []) when is_list(opts) do
     {index, document_id, opts} = extract_required_index_and_required_id!(opts)
+    {source_only, opts} = Keyword.pop(opts, :source_only, false)
+    segment = if(source_only, do: :_source, else: :_doc)
 
-    Client.get("#{index}/_source/#{document_id}", nil, nil, opts)
+    Client.get("#{index}/#{segment}/#{document_id}", nil, nil, opts)
   end
 
   @doc """